test_vertex_array_index
